requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pykitti>=0.3.1
rosbags>=0.9
tqdm>=4.65
pandas>=2.0
pyyaml>=6.0
pytest>=7.4
